fix the generator command
